refactor: 구분자 파싱 메서드를 분리하여 클래스로 정의하여 사용
