+++ src/StringCalculator.js
@@ -1,8 +1,10 @@
 import StringValidator from "./StringValidator.js";
+import DelimiterParser from "./DelimiterParser.js";
 
 class StringCalculator {
   constructor() {
     this.stringValidator = new StringValidator();
+    this.delimiterParser = new DelimiterParser();
   }
 
   // 입력값을 받아 최종 결과를 반환하는 메서드
@@ -15,7 +17,7 @@ class StringCalculator {
     }
 
     const { numberStrings, customDelimiter } =
-      this.parseBasicAndCustomDelimiter(preProcessInput); // 기본 구분자와 커스텀 구분자를 파싱함.
+      this.delimiterParser.parse(preProcessInput); // 기본 구분자와 커스텀 구분자를 파싱함.
 
     this.stringValidator.validateMultipleDelimiterInput(
       preProcessInput,
@@ -42,26 +44,6 @@ class StringCalculator {
   // 문자열을 숫자로 변환하는 메서드
   parseNumbers(numberStrings) {
     return numberStrings.map((number) => parseInt(number, 10));
-  }
-
-  // 커스텀 구분자를 포함한 모든 구분자 파싱 메서드
-  parseBasicAndCustomDelimiter(input) {
-    const customDelimiter = /^\/\/(.)\n(.*)/; // (.)은 커스텀 구분자 사이에 있는 하나의 문자를 의미하며 (.*)은 커스텀 구분자 이후의 문자열을 의미한다.
-    const match = input.match(customDelimiter);
-
-    // 커스텀 구분자가 있는 경우
-    if (match) {
-      const [, delimiter, numbers] = match;
-      return {
-        numberStrings: numbers.split(new RegExp(`[${delimiter},:]`)),
-        customDelimiter: delimiter,
-      };
-    }
-
-    return {
-      numberStrings: input.split(/[,:]/),
-      customDelimiter: null,
-    };
   }
 
   // 숫자 배열을 받아 합을 구하는 메서드
